This changeset implements a more robust configuration system for building h5py.
Currently h5py cannot be installed with pip on a Debian system because
the hdf5 library is not named libhdf5.so, due to Debian providing
multiple version of the library (serial and mpi-enabled), and cannot be
found at the place where the current configuration utility expect the
file to be found, due to Debian being multiarch.

This patch splits the autodection process into seperate steps for the
library itself (path and name of the library seperately), the header
files, and the version.

For this the rewrite provides seperate functions
autodetect_libdirs
autodetect_libname
autodetect_incluedirs
autodetect_version

All function should fall back to the current situation if they can not
contribute any additional information.

autodetect_libdirs tries to detect the library path using the ldconfig
tool and falls back to te pkgconfig tool, which works sometimes, but
e.g. under Debian is unable to provide information for the mpi-enabled
library version (if both are installed)

autodetect_includedirs falls back to the pkgconfig information as well,
but has some hardwired information for where to find the header files on
a Debian system.

autodetect_libname now holds the logic for how the hdf5 library should
be named and produces the appropriate regexp for autodetect_version and
how the library should be named when including it in the build_process.

All function have been amended to take as many additional parameters
during install time as possible for this the following commandline
switches have been introduced:

--hdf5-librdir
--hdf5-libname
--hdf5-includedir

all of them can also be accessed via the environment variables:

HDF5_LIB
HDF5_LIBNAME
HDF5_INCLUDE

additionaly an environment variable for the --mpi switch as been
introduced.

Conflicts:
	setup_configure.py