This changeset implements a more robust configuration system for building h5py.
h5py cannot be installed with pip on a Debian system because libhdf5 is not named
libhdf5 and is in in very non-standard places thus breaking the configure step
and the built.

This patch splits the autodection process into seperate steps for library,
header files, version and naming of the libraries itself.